Don't count edge touching as intersecting
Changed the < and > signs in intersects() to <= and >=. Since this is in
a not operation, this doesn't count the situation when, e.g. other.x1 ==
this.x2 (x edges touch).

This was a small problem in collision detection and with this, we can
avoid collision detection when they are just touching.